Coupons available History && Coupons Claiming && Search service

+++ lib/models/api/coupon_model.dart
@@ -23,7 +23,7 @@ class CouponModel {
   int couponId;
   int vid;
   String couponCode;
-  int percentageOff;
+  double percentageOff;
   bool singleUse;
   bool featureCoupon;
   DateTime startDate;
